add fn resetMap
