+++ js/main.js
@@ -1,6 +1,6 @@
 import './mock.js';
 import './popup.js';
-import './form.js';
 import './validity.js';
 import './map.js';
+import './form.js';
 import './api.js';
