+++ js/map.js
@@ -1,4 +1,4 @@
-import {activatePage} from './form.js';
+import {activatePage, resetForm} from './form.js';
 import {createCustomPopup} from './popup.js';
 
 const DEFAULT_COORDINATES = {
@@ -11,8 +11,6 @@ const SECONDARY_ICON_SIZE = [40, 40];
 const SECONDARY_ICON_ANCHOR = [20, 40];
 
 const addressInput = document.querySelector('#address');
-const offerForm = document.querySelector('.ad-form');
-const filterForm = document.querySelector('.map__filters');
 const resetButton = document.querySelector('.ad-form__reset');
 
 const map = L.map('map-canvas')
@@ -91,13 +89,7 @@ const renderSimilarList = (similarOffers) => {
   });
 };
 
-const resetForm = () => {
-  offerForm.reset();
-  filterForm.reset();
-  document.querySelector('#price').placeholder = 1000;
-  document.querySelector('#price').min = 0;
-  addressInput.value = `${DEFAULT_COORDINATES.lat}, ${DEFAULT_COORDINATES.lng}`;
-
+const resetMap = () => {
   map
     .setView({
       lat: DEFAULT_COORDINATES.lat,
@@ -117,4 +109,4 @@ resetButton.addEventListener('click', (evt) => {
   resetForm();
 });
 
-export {renderSimilarList, resetForm};
+export {DEFAULT_COORDINATES, renderSimilarList, resetMap};
